requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.10.0
django-extensions==3.1.3
django-filter==21.1
django-restql==0.15.1
djangorestframework==3.12.4
Markdown==3.3.4
psycopg2==2.9.1
PyJWT==1.7.1
pyPEG2==2.15.2
pytz==2021.1
sqlparse==0.4.2
